Move MainLayout to a Layouts folder
